+++ install_ydk.sh
@@ -57,6 +57,10 @@ Options and arguments:
 Environment variables:
 YDKGEN_HOME         specifies location of ydk-gen git repository;
                     if not set, \$HOME/ydk-gen is assumed
+PYTHON_BIN          specifies alternative location of Python executable;
+                    if not set, python3 is assumed
+PIP_BIN             specifies alternative location of Pip executable;
+                    if not set, pip3 is assumed
 PYTHON_VENV         specifies location of python virtual environment;
                     if not set, \$HOME/venv is assumed
 GOROOT              specifies installation directory of go software;
@@ -72,44 +76,63 @@ CMAKE_LIBRARY_PATH  Location of Python shared libraries;
 }
 
 function activate_python_venv() {
-  if [[ -z ${PYTHON_VENV} ]]; then
+  if [ -z ${PYTHON_VENV} ]; then
     PYTHON_VENV=${HOME}/venv
     print_msg "Python virtual environment location is set to ${PYTHON_VENV}"
   fi
   if [[ ! -d ${PYTHON_VENV} ]]; then
     print_msg "Creating Python3 virtual environment in ${PYTHON_VENV}"
-    run_cmd python3 -m venv "${PYTHON_VENV}"
+    run_cmd $PYTHON_BIN -m venv "${PYTHON_VENV}"
   fi
   run_cmd source "${PYTHON_VENV}"/bin/activate
 }
 
 function init_py_env {
+  python_version=$($PYTHON_BIN -V | awk '{ print $2 }' | cut -d '.' -f 2)
+  if [[ $python_version -gt 11 ]]
+  then
+    print_msg "YDK does not support Python 3.$python_version"
+    python3x_location=$(find /usr/local/bin/ -name 'python3.11' | tail -n 1)
+    if [ -z $python3x_location ]
+    then
+      print_msg "Python3.11 local installation is not found; exiting"
+      exit 1
+    else
+      export PYTHON_BIN=python3.11
+      export PIP_BIN=pip3.11
+      if [ $install_venv == "no" ]; then
+        print_msg "Installation of virtual environment for non-system Python installation is enforced"
+        install_venv=yes
+      fi
+    fi
+  fi
+
   if [ $install_venv == "yes" ]; then
     activate_python_venv
   fi
   print_msg "Checking python3 version and installation"
-  python3 --version
+  $PYTHON_BIN --version
   status=$?
   if [ $status -ne 0 ]; then
     MSG_COLOR=$RED
-    print_msg "Could not locate python3 interpreter"
+    print_msg "Could not locate $PYTHON_BIN interpreter"
     exit $status
   fi
 
-  print_msg "Checking pip3 version and installation"
-  pip3 -V
+  print_msg "Checking $PIP_BIN version and installation"
+  $PIP_BIN --version
   status=$?
   if [ $status -ne 0 ]; then
     MSG_COLOR=$RED
-    print_msg "Could not locate pip3"
+    print_msg "Could not locate $PIP_BIN"
     exit $status
   fi
   export PIP_DISABLE_PIP_VERSION_CHECK=1
 
   if [[ $ydk_lang == "py" || $ydk_lang == "all" ]]; then
     print_msg "Checking installation of python shared libraries"
-    ver=$(python3 -c "import sys;print('{}.{}'.format(sys.version_info.major,sys.version_info.minor))")
-    ver_minor=$(python3 -c "import sys;print('{}'.format(sys.version_info.minor))")
+    ver=$($PYTHON_BIN -c "import sys;print('{}.{}'.format(sys.version_info.major,sys.version_info.minor))")
+    ver_minor=$($PYTHON_BIN -c "import sys;print('{}'.format(sys.version_info.minor))")
     os_type=$(uname)
     if [[ ${os_type} == "Linux" ]]; then
       if [ $ver_minor -gt 7 ]; then
@@ -120,10 +143,10 @@ function init_py_env {
     elif [[ ${os_type} == "Darwin" ]]; then
       ext="$ver.dylib"
     fi
-    lines=($(locate libpython$ext))
-    if [[ ${#lines[@]} -gt 0 ]] && [[ -x ${lines[0]} || -L ${lines[0]} ]]; then
-      if [[ -z $CMAKE_LIBRARY_PATH ]]; then
-        export CMAKE_LIBRARY_PATH=$(dirname ${lines[0]})
+    libpython_path=$(locate libpython$ext | head -n 1)
+    if [[ -n ${libpython_path} ]] && [[ -x ${libpython_path} || -L ${libpython_path} ]]; then
+      if [ -z $CMAKE_LIBRARY_PATH ]; then
+        export CMAKE_LIBRARY_PATH=$(dirname ${libpython_path})
         print_msg "Setting CMAKE_LIBRARY_PATH to $CMAKE_LIBRARY_PATH"
       fi
     else
@@ -133,7 +156,14 @@ function init_py_env {
     fi
   fi
   print_msg "Checking and installing Python requirements"
-  pip3 install -r requirements.txt
+  $PIP_BIN install pyang
+  status=$?
+  if [ $status -ne 0 ]; then
+    print_msg "Enabling sudo for Python components installation"
+    sudo_cmd="sudo"
+    sudo_flag="s"
+  fi
+  $sudo_cmd $PIP_BIN install -r requirements.txt
 }
 
 function init_go_env {
@@ -154,7 +184,7 @@ function init_go_env {
         fi
         export PATH=$GOROOT/bin:$PATH
     fi
-    if [[ -z $GOPATH ]]; then
+    if [ -z $GOPATH ]; then
         export GOPATH=$HOME/go
         mkdir -p $GOPATH
         print_msg "Setting GOPATH to $GOPATH"
@@ -166,6 +196,9 @@ function init_go_env {
 
     if [ ! -d $GOPATH/src/github.com/stretchr/testify ]; then
         go get github.com/stretchr/testify
+        go get github.com/davecgh/go-spew/spew
+        go get github.com/pmezard/go-difflib/difflib
+        go get gopkg.in/yaml.v3
         cd $GOPATH/src/github.com/stretchr/testify
         git checkout tags/v1.6.1
         cd -
@@ -181,7 +214,7 @@ function install_cpp_core {
     cd $YDKGEN_HOME
     local cpp_sudo_flag
     if [ $(id -u -n) != "root" ]; then cpp_sudo_flag="s"; fi
-    run_cmd python3 generate.py --core --cpp -i$cpp_sudo_flag
+    run_cmd $PYTHON_BIN generate.py --core --cpp -i$cpp_sudo_flag
 }
 
 function install_cpp_gnmi {
@@ -189,28 +222,28 @@ function install_cpp_gnmi {
     cd $YDKGEN_HOME
     local cpp_sudo_flag
     if [ $(id -u -n) != "root" ]; then cpp_sudo_flag="s"; fi
-    run_cmd python3 generate.py --service profiles/services/gnmi-0.4.0.json --cpp -i$cpp_sudo_flag
+    run_cmd $PYTHON_BIN generate.py --service profiles/services/gnmi-0.4.0.json --cpp -i$cpp_sudo_flag
 }
 
 function install_go_core {
     print_msg "Installing Go core packages"
     cd $YDKGEN_HOME
-    run_cmd python3 generate.py -i --core --go
+    run_cmd $PYTHON_BIN generate.py -i --core --go
 }
 
 function install_go_gnmi {
     print_msg "Installing Go gNMI service package"
     cd $YDKGEN_HOME
-    run_cmd python3 generate.py -i --service profiles/services/gnmi-0.4.0.json --go
+    run_cmd $PYTHON_BIN generate.py -i --service profiles/services/gnmi-0.4.0.json --go
 }
 
 function install_py_core {
     print_msg "Building and installing Python core package"
     cd $YDKGEN_HOME
-    run_cmd python3 generate.py -i --core
+    run_cmd $PYTHON_BIN generate.py -i$sudo_flag --core
 
     print_msg "Verifying Python YDK core package installation"
-    python3 -c "from ydk.path import NetconfSession"
+    $PYTHON_BIN -c "from ydk.path import NetconfSession"
     local status=$?
     if [ $status -ne 0 ]; then
         MSG_COLOR=$RED
@@ -222,10 +255,10 @@ function install_py_core {
 function install_py_gnmi {
     print_msg "Installing gNMI package for Python"
     cd $YDKGEN_HOME
-    run_cmd python3 generate.py -i --service profiles/services/gnmi-0.4.0.json
+    run_cmd $PYTHON_BIN generate.py -i$sudo_flag --service profiles/services/gnmi-0.4.0.json
 
     print_msg "Verifying Python gNMI package installation"
-    python3 -c "from ydk.gnmi.path import gNMISession"
+    $PYTHON_BIN -c "from ydk.gnmi.path import gNMISession"
     local status=$?
     if [ $status -ne 0 ]; then
         MSG_COLOR=$RED
@@ -328,21 +361,21 @@ export YDKGEN_HOME=\"${YDKGEN_HOME}\"
 export C_INCLUDE_PATH=\"$C_INCLUDE_PATH\"
 export CPLUS_INCLUDE_PATH=\"$CPLUS_INCLUDE_PATH\"
 export PIP_DISABLE_PIP_VERSION_CHECK=1
-alias python=python3
-alias pip=pip3
+alias python=$PYTHON_BIN
+alias pip=$PIP_BIN
 " > .env
-  if [[ -n $LD_LIBRARY_PATH ]]; then
+  if [ -n $LD_LIBRARY_PATH ]; then
     echo "export LD_LIBRARY_PATH=$LD_LIBRARY_PATH
 " >> .env
   fi
   if [ $install_venv == "yes" ]; then
     echo "export PYTHON_VENV=$PYTHON_VENV" >> .env
     echo "source $PYTHON_VENV/bin/activate" >> .env
-  elif [[ -n $python_location ]]; then
+  elif [ -n $python_location ]; then
     echo "export PATH=$python_location/bin:\$PATH
 " >> .env
   fi
-  if [[ -n $sudo_cmd ]]; then
+  if [ -n $sudo_cmd ]; then
     echo "export SUDO_CMD=$sudo_cmd" >> .env
   fi
   if [[ $ydk_lang == "go" || $ydk_lang == "all" ]]; then
@@ -369,7 +402,7 @@ if [[ \$go_version > \"1.11.\" ]]; then
 fi
 " >> .env
   fi
-  if [[ -n $CMAKE_LIBRARY_PATH ]]; then
+  if [ -n $CMAKE_LIBRARY_PATH ]; then
     echo "export CMAKE_LIBRARY_PATH=$CMAKE_LIBRARY_PATH
 " >> .env
   fi
@@ -395,6 +428,8 @@ dependencies="yes"
 install_venv="no"
 sudo_flag=
 sudo_cmd=
+[ -z "${PYTHON_BIN}" ] && PYTHON_BIN=python3
+[ -z "${PIP_BIN}" ] && PIP_BIN=pip3
 
 # As long as there is at least one more argument, keep looping
 while [[ $# -gt 0 ]]; do
@@ -469,7 +504,7 @@ libydk_path="libydk-$ydk_version.a"
 libydk_gnmi_path="libydk_gnmi-$gnmi_version.a"
 echo "YDK-$ydk_version installation options:"
 if [ ${install_venv} == "no" ]; then
-  if [[ -n $python_location ]]; then
+  if [ -n $python_location ]; then
     echo " - use custom Python installation in ${python_location}"
   else
     echo " - use system Python installation"
@@ -503,7 +538,7 @@ print_msg "Running OS type: $os_type"
 print_msg "OS info: $os_info"
 if [[ ${os_type} == "Linux" ]]; then
   if [[ ${os_info} == *"Ubuntu"* ]]; then
-    if [[ ${os_info} != *"xenial"* && ${os_info} != *"bionic"* && ${os_info} != *"focal"* && ${os_info} != *"jammy"* ]]; then
+    if [[ ${os_info} != *"xenial"* && ${os_info} != *"bionic"* && ${os_info} != *"focal"* && ${os_info} != *"jammy"* && ${os_info} != *"noble"* ]]; then
         print_msg "WARNING! Unsupported Ubuntu distribution found. Will try the best efforts."
     fi
   elif [[ ${os_info} == *"fedora"* ]]; then
@@ -524,19 +559,28 @@ if [[ ${os_type} == "Linux" ]]; then
   fi
 fi
 
-if [[ -z ${C_INCLUDE_PATH} ]]; then
+if [ -z ${C_INCLUDE_PATH} ]; then
     export C_INCLUDE_PATH=/usr/local/include
 fi
-if [[ -z ${CPLUS_INCLUDE_PATH} ]]; then
+if [ -z ${CPLUS_INCLUDE_PATH} ]; then
     export CPLUS_INCLUDE_PATH=/usr/local/include
+    if [ $(uname) == "Linux" ]; then
+      gcc_version=$(gcc --version | head -n 1 | awk '{ print $3 }')
+      major=$(echo $gcc_version | cut -d '.' -f 1)
+      export CPLUS_INCLUDE_PATH=/usr/include/c++/$major:$CPLUS_INCLUDE_PATH
+    fi
+    print_msg "CPLUS_INCLUDE_PATH is set to: $CPLUS_INCLUDE_PATH"
 fi
 
-if [[ ${os_info} == *"fedora"* || ${os_info} == *"jammy"* ]]; then
+if [[ ${os_info} == *"fedora"* || ${os_info} == *"jammy"* || ${os_info} == *"noble"* ]]; then
   if [[ $LD_LIBRARY_PATH != *"/usr/local/lib"* ]]; then
     export LD_LIBRARY_PATH="/usr/local/lib:$LD_LIBRARY_PATH"
   fi
   if [[ $LD_LIBRARY_PATH != *"/lib64"* && -d /usr/lib64 ]]; then
-    export LD_LIBRARY_PATH=$LD_LIBRARY_PATH:/usr/local/lib64:/usr/lib64
+    export LD_LIBRARY_PATH="$LD_LIBRARY_PATH:/usr/local/lib64:/usr/lib64"
+  fi
+  if [[ $LD_LIBRARY_PATH != *"/usr/local/ssl/lib"* && -d /usr/local/ssl/lib ]]; then
+    export LD_LIBRARY_PATH="/usr/local/ssl/lib:$LD_LIBRARY_PATH"
   fi
   print_msg "LD_LIBRARY_PATH is set to: $LD_LIBRARY_PATH"
 fi