+++ test/tests.sh
@@ -244,6 +244,9 @@ function init_go_env {
 
     if [ ! -d $GOPATH/src/github.com/stretchr/testify ]; then
         go get github.com/stretchr/testify
+        go get github.com/davecgh/go-spew/spew
+        go get github.com/pmezard/go-difflib/difflib
+        go get gopkg.in/yaml.v3
         cd $GOPATH/src/github.com/stretchr/testify
         git checkout tags/v1.6.1
         cd -